refactor: rename useTokenOrUpdate to useTokenOrRefresh
